Rename backtest to analytics-engine
